refactor: create a new `index.ts` to expose our public APIs
