+++ src/cli.ts
@@ -1,7 +1,7 @@
 import { info, hostInfo, warning } from './common';
 import fs from 'fs-extra';
 import { initializeProxy } from '@electron/get';
-import packager from '..';
+import { packager } from './packager';
 import path from 'path';
 import yargs from 'yargs-parser';
