+++ src/index.ts
@@ -1,0 +1,7 @@
+import { packager } from './packager';
+
+export default packager;
+
+export { packager };
+
+export * from './types';
